Add support for verifying types using mypy's unit testing
This adds support for verifying that certain code type checks
and that certain code doesn't. This allows us to verify
that are typings are behaving as they should.

It does this by re-using MyPy's builtin testing framework
which is documented here:
https://github.com/python/mypy/blob/master/test-data/unit/README.md

The only other project I found which does this is the SQLAlchemy stubs:

https://github.com/angaza/sqlalchemy-stubs/blob/6af1f077745e91d8a4ca624e19cdcf7082aaabdf/test/testsql.py

It ops for another approach however. Instead of reusing the TypeChecker
test from MyPy, it builds a new test case which calls out to
the MyPy CLI.
I choose instead to re-use that test case, so that hopefully it would
stay up to date more cleanly.