Perform CSE when converting to egglog expression
Currently, when converting a tree of Python expressions to eggglog,
pieces that are duplicates will be included twice. This is correct,
but produced a very large expression statement, compared to if we did CSE.

This does that by assigning a let binding to all call expressions as we convert them to egglog.