Remove automatic extraction in builtin evaluation
In https://github.com/egraphs-good/egglog-python/pull/265 (released as
9.0.0) the ability to automatically extract a builtin using a global
egraph context was added. This PR removes that feature, requiring all
builtins to be in a normalized form.

I realized that for https://github.com/egraphs-good/egglog-python/issues/241
we want facts to compare structural equality when converting to a boolean,
instead of using the e-graph. Looking at that previous PR, it seems like
a mistake to add this implicit context, making things more confusing
and opaque with minimal UX improvements.