Add support for `set_cost` action to have row level costs for extraction
This re-implements the syntactic macro in the egglog experimental to
create a new table for each function to store its cost. It re-uses the
custom extractor which looks up in this table.

I talked to @yihongz and we discussed different ways to implement this,
from changing syntactic macros to take Expr instead of Sexpr, but settled
on just having me re-implement the syntax.